feat(api): implement `NodeStringHasher`

+++ src/server/modules/identity/application/use-cases/DeleteUser.ts
@@ -3,6 +3,7 @@ import { UUID } from "@/server/utils";
 import { User } from "../../domain/entities/User";
 import { UserRepository } from "../../domain/repositories/UserRepository";
 import { DeleteUserDTO } from "../dto/DeleteUserDTO";
+import { UserNotFoundError } from "../errors/UserNotFoundError";
 
 export class DeleteUser {
     constructor(private readonly userRepository: UserRepository) {}
@@ -12,6 +13,9 @@ export class DeleteUser {
 
         const userId = UUID.fromString(dto.id);
         const user = await this.userRepository.getById(userId);
+
+        if (!user) throw new UserNotFoundError("UUID", dto.id);
+
         await this.userRepository.delete(user);
     }
 }
